+++ src/RateProviders/NbuRateProvider.php
@@ -2,11 +2,18 @@
 
 namespace Fomvasss\Currency\RateProviders;
 
-class NbuRateProvider extends AbstractRateProvider
+use Fomvasss\Currency\Contracts\HistoricalRateProvider;
+
+class NbuRateProvider extends AbstractRateProvider implements HistoricalRateProvider
 {
     protected function getApiUrl(): string
     {
         return 'https://bank.gov.ua/NBUStatService/v1/statdirectory/exchange?json';
+    }
+
+    protected function getHistoricalApiUrl(\DateTimeInterface $date): string
+    {
+        return $this->getApiUrl() . '&date=' . $date->format('Ymd');
     }
 
     protected function parseResponse($response): array